+++ HARK/Calibration/tests/test_load_data.py
@@ -1,8 +1,10 @@
 import unittest
 
-from HARK.datasets import load_SCF_wealth_weights
-from HARK.datasets.cpi.us.CPITools import cpi_deflator
-from HARK.datasets.SCF.WealthIncomeDist.SCFDistTools import income_wealth_dists_from_scf
+from HARK.Calibration import load_SCF_wealth_weights
+from HARK.Calibration.cpi.us.CPITools import cpi_deflator
+from HARK.Calibration.SCF.WealthIncomeDist.SCFDistTools import (
+    income_wealth_dists_from_scf,
+)
 
 
 class test_load_SCF_wealth_weights(unittest.TestCase):
