+++ HARK/ConsumptionSaving/ConsIndShockModel.py
@@ -26,8 +26,10 @@ from HARK.Calibration.Income.IncomeProcesses import (
     get_PermShkDstn_from_IncShkDstn,
     get_TranShkDstn_from_IncShkDstn,
 )
-from HARK.datasets.life_tables.us_ssa.SSATools import parse_ssa_life_table
-from HARK.datasets.SCF.WealthIncomeDist.SCFDistTools import income_wealth_dists_from_scf
+from HARK.Calibration.life_tables.us_ssa.SSATools import parse_ssa_life_table
+from HARK.Calibration.SCF.WealthIncomeDist.SCFDistTools import (
+    income_wealth_dists_from_scf,
+)
 from HARK.distribution import (
     Lognormal,
     MeanOneLogNormal,
