--- HARK/datasets/__init__.py
@@ -1,1 +1,0 @@
-from HARK.datasets.load_data import *
